// IntelliPMS-Frontdesk/src/app/models/employee.model.ts
export class EmployeeModel {
  employeeId: string;
  employeeRole: string;
  employeeName: string;
  gender: string;
  firstName: string;
  lastName: string;
  otherName: string;
  emailAddress: string;
  phoneNumber: number;
  operateRegister: string;
  address: string;
  picture: string;
  statuss: string;
}